+++ preserve-ui/src/api/jellyfin/types.ts
@@ -1,6 +1,9 @@
 import { BaseServerDefinition } from '../interface';
 export const JELLYFIN_SERVER_TYPE = 'jellyfin';
 
+export const TRACK_PIMARY_ART_FIELD = 'primaryArtTag';
+export const ALBUM_PRIMARY_ART_FIELD = 'albumPrimaryArtTag';
+
 export enum ConnectionMode {
     Local = 0,
     Remote = 1,